requirements.txt
fastapi
uvicorn[standard]
websockets
numpy
opencv-python
ultralytics
google-generativeai
gtts
pillow
python-multipart